More cleanup to split cil.c from ESIL code
