+++ librz/arch/isa/avr/avr_esil.h
@@ -5,7 +5,7 @@
 #define RZ_AVR_ESIL_H
 #include "analysis_private.h"
 
-RZ_IPI bool rz_avr_esil_init(RzAnalysisEsil *esil);
+RZ_IPI bool rz_avr_esil_init(void *pesil);
 RZ_IPI void rz_avr_esil_opcode(RzAnalysis *analysis, RzAnalysisOp *op, ut64 addr, const ut8 *buf, int len);
 
 #endif /* RZ_AVR_ESIL_H */
