+++ librz/core/tui/esil.c
@@ -49,8 +49,8 @@ RZ_IPI bool rz_core_visual_esil(RzCore *core) {
 	}
 	memcpy(buf, core->block, sizeof(ut64));
 	RzAnalysisEsil *esil = rz_analysis_esil_new(20, 0, addrsize);
-	esil->core = core;
-	esil->analysis = core->analysis;
+	esil->pcore = core;
+	esil->panalysis = core->analysis;
 	rz_analysis_esil_set_pc(esil, core->offset);
 	for (;;) {
 		rz_cons_clear00();
