+++ librz/include/rz_esil/rz_esil.h
@@ -6,17 +6,11 @@
 #define RZ_ANALYSIS_ESIL_H
 
 #include <rz_util.h>
-#include <rz_bind.h>
-#include <rz_lib.h>
 #include <rz_reg.h>
 
 #ifdef __cplusplus
 extern "C" {
 #endif
-
-typedef struct rz_analysis_t RzAnalysis;
-typedef struct rz_analysis_op_t RzAnalysisOp;
-typedef struct rz_il_trace_instruction_t RzILTraceInstruction;
 
 #define RZ_ANALYSIS_ESIL_GOTO_LIMIT 4096
 
@@ -127,8 +121,8 @@ typedef struct rz_analysis_esil_inter_state_t {
 } RzAnalysisEsilInterState;
 
 struct rz_analysis_esil_t {
-	RzCore *core;
-	RzAnalysis *analysis;
+	void *pcore;
+	void *panalysis;
 	char **stack;
 	ut64 addrmask;
 	int stacksize;
@@ -240,7 +234,7 @@ typedef struct rz_analysis_esil_memory_region_t {
 
 RZ_API RzAnalysisEsil *rz_analysis_esil_new(int stacksize, int iotrap, unsigned int addrsize);
 RZ_API bool rz_analysis_esil_set_pc(RzAnalysisEsil *esil, ut64 addr);
-RZ_API bool rz_analysis_esil_setup(RzAnalysisEsil *esil, RzAnalysis *analysis, int romem, int stats, int nonull, RzCore *core);
+RZ_API bool rz_analysis_esil_setup(RzAnalysisEsil *esil, void /*RzAnalysis*/ *analysis, int romem, int stats, int nonull, void /*RzCore*/ *core);
 RZ_API void rz_analysis_esil_free(RzAnalysisEsil *esil);
 RZ_API bool rz_analysis_esil_runword(RzAnalysisEsil *esil, const char *word);
 RZ_API bool rz_analysis_esil_parse(RzAnalysisEsil *esil, const char *str);
@@ -272,10 +266,10 @@ RZ_API void rz_analysis_esil_mem_ro(RzAnalysisEsil *esil, int mem_readonly);
 RZ_API void rz_analysis_esil_stats(RzAnalysisEsil *esil, int enable);
 
 /* ESIL trace */
-RZ_API RZ_BORROW RzILTraceInstruction *rz_analysis_esil_get_instruction_trace(RZ_NONNULL RzAnalysisEsilTrace *etrace, int idx);
+RZ_API RZ_BORROW void /*RzILTraceInstruction*/ *rz_analysis_esil_get_instruction_trace(RZ_NONNULL RzAnalysisEsilTrace *etrace, int idx);
 RZ_API RzAnalysisEsilTrace *rz_analysis_esil_trace_new(RzAnalysisEsil *esil);
 RZ_API void rz_analysis_esil_trace_free(RzAnalysisEsilTrace *trace);
-RZ_API void rz_analysis_esil_trace_op(RzAnalysisEsil *esil, RZ_NONNULL RzAnalysisOp *op);
+RZ_API void rz_analysis_esil_trace_op(RzAnalysisEsil *esil, ut64 pc, RZ_NULLABLE const char *esil_expr);
 RZ_API void rz_analysis_esil_trace_list(RzAnalysisEsil *esil);
 RZ_API void rz_analysis_esil_trace_show(RzAnalysisEsil *esil, int idx);
 RZ_API void rz_analysis_esil_trace_restore(RzAnalysisEsil *esil, int idx);
